I added a float on line 14

+++ Cilindro/Program.cs
@@ -11,7 +11,7 @@ namespace Cilindro
             Console.Write("Enter the radius of the cylinder: ");
             float radius = float.Parse(Console.ReadLine());
 
-            float V = Math.PI * Math.Pow(radius, 2) * height;
+            float V = Math.PI * (float)Math.Pow(radius, 2) * height;
             float Sa = 2 * Math.PI * radius * (radius + height);
 
             Console.WriteLine($"The volume of the cylinder {V:F2}");
